Fold .testing.client.Client into .testing.Harness
