+++ aspen/testing/__init__.py
@@ -12,7 +12,7 @@ import sys
 from collections import namedtuple
 
 from .. import resources
-from .client import Client
+from ..website import Website
 from filesystem_tree import FilesystemTree
 
 
@@ -44,11 +44,22 @@ class Harness(object):
         self.fs = namedtuple('fs', 'www project')
         self.fs.www = FilesystemTree()
         self.fs.project = FilesystemTree()
-        self.client = Client(self.fs.www.root, self.fs.project.root)
+        self._website = None
 
     def teardown(self):
         self.fs.www.remove()
         self.fs.project.remove()
+
+    def hydrate_website(self, **kwargs):
+        if (self._website is None) or kwargs:
+            _kwargs = { 'www_root': self.fs.www.root
+                      , 'project_root': self.fs.project.root
+                       }
+            _kwargs.update(kwargs)
+            self._website = Website(**_kwargs)
+        return self._website
+
+    website = property(hydrate_website)
 
 
     # Simple API
@@ -61,7 +72,7 @@ class Harness(object):
         if filepath is not None:
             self.fs.www.mk((filepath, contents))
         if website_configuration is not None:
-            self.client.hydrate_website(**website_configuration)
+            self.hydrate_website(**website_configuration)
 
         if uripath is None:
             if filepath is None:
@@ -70,12 +81,37 @@ class Harness(object):
                 uripath = '/' + filepath
                 if uripath.endswith('.spt'):
                     uripath = uripath[:-len('.spt')]
-                for indexname in self.client.website.indices:
+                for indexname in self.website.indices:
                     if uripath.endswith(indexname):
                         uripath = uripath[:-len(indexname)]
                         break
 
-        return self.client._hit('GET', uripath, querystring, **kw)
+        return self._hit('GET', uripath, querystring, **kw)
+
+    def _hit(self, method, path='/', querystring='', raise_immediately=True, return_after=None,
+             want='response', **headers):
+
+        state = self.website.respond( path
+                                    , querystring
+                                    , accept_header=None
+                                    , raise_immediately=raise_immediately
+                                    , return_after=return_after
+                                     )
+
+        response = state.get('response')
+        if response is not None:
+            if response.headers.cookie:
+                self.cookie.update(response.headers.cookie)
+
+        attr_path = want.split('.')
+        base = attr_path[0]
+        attr_path = attr_path[1:]
+
+        out = state[base]
+        for name in attr_path:
+            out = getattr(out, name)
+
+        return out
 
     def make_dispatch_result(self, *a, **kw):
         kw['return_after'] = 'dispatch_path_to_filesystem'
