+++ tests/test_json_resource.py
@@ -26,7 +26,7 @@ def test_json_defaults_to_application_json_for_static_json(harness):
     assert actual == 'application/json'
 
 def test_json_content_type_is_configurable_for_static_json(harness):
-    harness.client.website.media_type_json = "floober/blah"
+    harness.website.media_type_json = "floober/blah"
     expected = 'floober/blah'
     actual = harness.simple( '{"Greetings": "program!"}'
                            , filepath="foo.json"
@@ -41,7 +41,7 @@ def test_json_content_type_is_configurable_from_kwargs(harness):
     assert actual == 'floober/blah'
 
 def test_json_content_type_is_configurable_for_dynamic_json(harness):
-    harness.client.website.media_type_json = "floober/blah"
+    harness.website.media_type_json = "floober/blah"
     actual = harness.simple( "[---]\n[---] floober/blah\n{'Greetings': 'program!'}"
                            , filepath="foo.json.spt"
                             ).headers['Content-Type']
